requirements.txt
backend_service @ git+https://github.com/Sheesh1006/proto.git@6926fb502090a5fc76224358c6069f72519307be#subdirectory=gen/backend_service
chardet==5.2.0
chatgpt_service @ git+https://github.com/Sheesh1006/proto.git@d39873b57109a2940c5458ab0b528e72d92017cd#subdirectory=gen/chatgpt_service
grpcio==1.71.0
grpcio-tools==1.71.0
munch==4.0.0
pillow==11.2.1
protobuf==5.29.4
PyYAML==6.0.2
reportlab==4.4.0
setuptools==78.1.0
